refactor(runner): use schema for llm inputs and fix circular deps
- Refactor llm, generateText, generateObject to use JSON schemas for inputs and single result cells.
- Extract shared LLM schemas to packages/runner/src/builtins/llm-schemas.ts to resolve circular dependency.
- Fix generateObject tests to use valid dummy tools and force tool-calling path.
